requirements/pytorch_req.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.2=py38hecd8cb5_1001
archspec=0.2.3=pyhd3eb1b0_0
asttokens=2.0.5=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
boltons=23.0.0=py38hecd8cb5_0
bottleneck=1.4.0=py38h63504b1_0
brotli-python=1.0.9=py38hcec6c5f_8
bzip2=1.0.8=h6c40b1e_6
c-ares=1.19.1=h6c40b1e_0
ca-certificates=2024.6.2=h8857fd0_0
certifi=2024.6.2=py38hecd8cb5_0
cffi=1.16.0=py38h6c40b1e_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
comm=0.2.1=py38hecd8cb5_0
conda=24.5.0=py38hecd8cb5_0
conda-libmamba-solver=24.1.0=pyhd3eb1b0_0
conda-package-handling=2.3.0=py38hecd8cb5_0
conda-package-streaming=0.10.0=py38hecd8cb5_0
cpuonly=2.0=0
cramjam=2.8.3=pypi_0
debugpy=1.6.7=py38hcec6c5f_0
decorator=5.1.1=pyhd3eb1b0_0
distro=1.9.0=py38hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
fastparquet=2024.2.0=pypi_0
ffmpeg=4.3=h0a44026_0
fftw=3.3.9=h6c40b1e_2
fmt=9.1.0=ha357a0b_1
freetype=2.12.1=hd8bbffd_0
frozendict=2.4.2=py38h6c40b1e_0
fsspec=2024.6.0=pypi_0
gettext=0.21.0=he85b6c0_1
gmp=6.2.1=he9d5cce_3
gnutls=3.6.15=hed9c0bf_0
greenlet=3.0.1=py38hcec6c5f_0
icu=73.1=hcec6c5f_0
idna=3.7=py38hecd8cb5_0
importlib-metadata=7.0.1=py38hecd8cb5_0
importlib_metadata=7.0.1=hd3eb1b0_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.28.0=py38hecd8cb5_0
ipython=8.12.2=py38hecd8cb5_0
jedi=0.18.1=py38hecd8cb5_1
jinja2=3.1.4=py38hecd8cb5_0
joblib=1.4.2=py38hecd8cb5_0
jpeg=9e=h6c40b1e_1
jsonpatch=1.33=py38hecd8cb5_1
jsonpointer=2.1=pyhd3eb1b0_0
jupyter_client=8.6.0=py38hecd8cb5_0
jupyter_core=5.7.2=py38hecd8cb5_0
krb5=1.20.1=h428f121_1
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libarchive=3.6.2=h29ab7a1_3
libcurl=8.7.1=hf20ceda_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.17=hb664fd8_1
libedit=3.1.20230828=h6c40b1e_0
libev=4.33=h9ed2024_1
libffi=3.4.4=hecd8cb5_1
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libiconv=1.16=h6c40b1e_3
libidn2=2.3.4=h6c40b1e_0
libmamba=1.5.8=h258a794_2
libmambapy=1.5.8=py38h8c3233a_2
libnghttp2=1.57.0=h9beae6a_0
libpng=1.6.39=h6c40b1e_0
libpq=12.17=h04015c4_0
libsodium=1.0.18=h1de35cc_0
libsolv=0.7.24=hfff2838_1
libssh2=1.11.0=hf20ceda_0
libtasn1=4.19.0=h6c40b1e_0
libtiff=4.5.1=hcec6c5f_0
libunistring=0.9.10=h9ed2024_0
libuv=1.44.2=h6c40b1e_0
libwebp-base=1.3.2=h6c40b1e_0
libxml2=2.10.4=h45904e2_2
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_1
markupsafe=2.1.3=py38h6c40b1e_0
matplotlib-inline=0.1.6=py38hecd8cb5_0
menuinst=2.1.1=py38hecd8cb5_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py38h9ed2024_0
mkl_fft=1.3.1=py38h4ab4a9b_0
mkl_random=1.2.2=py38hb2f4e1b_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=py38hecd8cb5_0
nettle=3.7.3=h230ac6f_1
networkx=3.1=py38hecd8cb5_0
numexpr=2.8.4=py38hdaefe69_0
numpy=1.24.3=py38he696674_0
numpy-base=1.24.3=py38h9cd3388_0
openh264=2.1.1=h8346a28_0
openjpeg=2.4.0=h66ea3da_0
openssl=3.3.1=h87427d6_0
packaging=23.2=py38hecd8cb5_0
pandas=2.0.3=py38h3ea8b11_0
parso=0.8.3=pyhd3eb1b0_0
pcre2=10.42=h9b97e30_1
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.3.0=py38h6c40b1e_0
pip=24.0=py38hecd8cb5_0
platformdirs=3.10.0=py38hecd8cb5_0
pluggy=1.0.0=py38hecd8cb5_1
prompt-toolkit=3.0.43=py38hecd8cb5_0
psutil=5.9.0=py38hca72f7f_0
psycopg2=2.9.9=py38h6c40b1e_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py4j=0.10.9.7=pypi_0
pybind11-abi=4=hd3eb1b0_1
pycosat=0.6.6=py38h6c40b1e_1
pycparser=2.21=pyhd3eb1b0_0
pyg=2.1.0=py38_torch_1.11.0_cpu
pygments=2.15.1=py38hecd8cb5_1
pyparsing=3.0.9=py38hecd8cb5_0
pysocks=1.7.1=py38_1
pyspark=3.5.1=pypi_0
python=3.8.19=h5ee71fb_0
python-dateutil=2.9.0post0=py38hecd8cb5_2
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.11.0=py3.8_0
pytorch-cluster=1.6.0=py38_torch_1.11.0_cpu
pytorch-mutex=1.0=cpu
pytorch-scatter=2.0.9=py38_torch_1.11.0_cpu
pytorch-sparse=0.6.15=py38_torch_1.11.0_cpu
pytz=2024.1=pyhd8ed1ab_0
pyzmq=25.1.2=py38hcec6c5f_0
readline=8.2=hca72f7f_0
reproc=14.2.4=hcec6c5f_2
reproc-cpp=14.2.4=hcec6c5f_2
requests=2.32.2=py38hecd8cb5_0
ruamel.yaml=0.17.21=py38hca72f7f_0
ruamel.yaml.clib=0.2.6=py38hca72f7f_1
scikit-learn=1.1.3=py38hcec6c5f_1
scipy=1.9.3=py38h3d31255_0
setuptools=69.5.1=py38hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlalchemy=2.0.31=py38hc718529_0
sqlite=3.45.3=h6c40b1e_0
stack_data=0.2.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.14=h4d00af3_0
torchaudio=0.11.0=py38_cpu
torchvision=0.12.0=py38_cpu
tornado=6.4.1=py38h46256e1_0
tqdm=4.66.4=py38h20db666_0
traitlets=5.14.3=py38hecd8cb5_0
typing-extensions=4.11.0=py38hecd8cb5_0
typing_extensions=4.11.0=py38hecd8cb5_0
urllib3=2.2.2=py38hecd8cb5_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.43.0=py38hecd8cb5_0
xz=5.4.6=h6c40b1e_1
yaml-cpp=0.8.0=hcec6c5f_1
zeromq=4.3.5=hcec6c5f_0
zipp=3.17.0=py38hecd8cb5_0
zlib=1.2.13=h4b97444_1
zstandard=0.22.0=py38h2d76c9a_0
zstd=1.5.5=hc035e20_2
